refactor: Add Month component to folder
