+++ src/assets/styles/global.module.css
@@ -14,6 +14,10 @@ html {
   font-size: 10px;
 }
 
+button {
+  cursor: pointer;
+}
+
 button:focus-visible {
   outline-width: 2px;
   outline-color: var(--whiteColor);
