+++ src/components/_plans/Calendar/components/MonthsList/MonthsList.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react';
 
-import { Month } from './components/Month';
+import { Month } from './components/Month/Month';
 
 import { useArrowFocus } from 'hooks/useArrowFocus';
